refactor(backend): move wrapToolResultPayload to shared contextIntegrations/
Move utility to contextIntegrations/toolResultFormatter.ts for reuse
across all context integrations. Simplified signature to (toolName, payload)
since all results are 'generated by backend'.
