+++ packages/backend/src/services/contextIntegrations/toolResultFormatter.ts
@@ -1,0 +1,34 @@
+/**
+ * @description: Shared utilities for formatting tool result payloads for orchestration.
+ * @footnote-scope: core
+ * @footnote-module: ToolResultFormatter
+ * @footnote-risk: low - Pure formatting utility, no side effects.
+ * @footnote-ethics: low - Formatting only, no impact on provenance or decisions.
+ */
+const MARKER_SEPARATOR = '// ==========';
+const MARKER_BEGIN = '// BEGIN Backend Tool Result';
+const MARKER_END = '// END Backend Tool Result';
+
+/**
+ * Wraps a JSON payload in standardized marker comments for orchestration.
+ * Used by context-step executors to format tool results for LLM consumption.
+ *
+ * @param toolName - The tool identifier (e.g., 'weather_forecast')
+ * @param payload - The JSON object to wrap
+ * @returns Formatted string with marker comments and JSON
+ */
+export const wrapToolResultPayload = (
+    toolName: string,
+    payload: object
+): string => {
+    return [
+        MARKER_SEPARATOR,
+        MARKER_BEGIN,
+        `// ${toolName} result - generated by backend`,
+        MARKER_SEPARATOR,
+        JSON.stringify(payload),
+        MARKER_SEPARATOR,
+        MARKER_END,
+        MARKER_SEPARATOR,
+    ].join('\n');
+};
