+++ packages/backend/src/services/openMeteoForecastTool.ts
@@ -19,35 +19,7 @@ const DEFAULT_HORIZON_PERIODS = 5;
 const MIN_HORIZON_PERIODS = 1;
 const MAX_HORIZON_PERIODS = 12;
 
-const MARKER_SEPARATOR = '// ==========';
-const MARKER_BEGIN = '// BEGIN Backend Tool Result';
-const MARKER_END = '// END Backend Tool Result';
-
-/**
- * Wraps a JSON payload in standardized marker comments for orchestration.
- * Used by context-step executors to format tool results for LLM consumption.
- *
- * @param toolName - The tool identifier (e.g., 'weather_forecast')
- * @param description - Brief description of what generated the result
- * @param payload - The JSON object to wrap
- * @returns Formatted string with marker comments and JSON
- */
-export const wrapToolResultPayload = (
-    toolName: string,
-    description: string,
-    payload: object
-): string => {
-    return [
-        MARKER_SEPARATOR,
-        MARKER_BEGIN,
-        `// ${toolName} result - ${description}`,
-        MARKER_SEPARATOR,
-        JSON.stringify(payload),
-        MARKER_SEPARATOR,
-        MARKER_END,
-        MARKER_SEPARATOR,
-    ].join('\n');
-};
+import { wrapToolResultPayload } from './contextIntegrations/toolResultFormatter.js';
 
 /**
  * Formats a WeatherForecastToolResult into a string payload for orchestration.
@@ -116,11 +88,7 @@ export const formatWeatherToolResultMessage = (
               }
             : result;
 
-    return wrapToolResultPayload(
-        result.toolName,
-        'generated by backend executor',
-        compactResult
-    );
+    return wrapToolResultPayload(result.toolName, compactResult);
 };
 
 type FetchLike = typeof fetch;
