+++ packages/backend/src/services/tools/weatherForecastToolAdapter.ts
@@ -14,7 +14,7 @@ import type {
     WeatherForecastTool,
     WeatherForecastToolResult,
 } from '../openMeteoForecastTool.js';
-import { wrapToolResultPayload } from '../openMeteoForecastTool.js';
+import { wrapToolResultPayload } from '../contextIntegrations/toolResultFormatter.js';
 import type { ChatGenerationWeatherRequest } from '../chatGenerationTypes.js';
 
 const formatWeatherToolResultMessage = (
@@ -45,11 +45,7 @@ const formatWeatherToolResultMessage = (
               }
             : result;
 
-    return wrapToolResultPayload(
-        result.toolName,
-        'generated by backend adapter',
-        compactResult
-    );
+    return wrapToolResultPayload(result.toolName, compactResult);
 };
 
 export const executeWeatherForecastTool = async ({
